# docs/swagger.yaml
definitions:
  common.State:
    properties:
      code:
        description: 状态码，0：成功；否则失败
        type: integer
      data:
        description: 数据，失败时返回空
      msg:
        description: 状态的具体描述
        type: string
    type: object
...
  model.MsgExtend:
    properties:
      remind:
        description: '@ 的描述信息'
        items:
          $ref: '#/definitions/model.Remind'
        type: array
    type: object
  model.Remind:
    properties:
      account_id:
        description: 被 @ 的账号 ID
        minimum: 1
        type: integer
      idx:
        description: 第几个 @
        minimum: 1
        type: integer
    required:
    - account_id
    - idx
    type: object
...
  reply.ParamBriefMsgInfo:
    properties:
      account_id:
        description: 账号ID 发送者ID
        type: integer
      create_at:
        description: 创建时间
        type: string
      file_id:
        description: 文件ID 当消息类型为file时>0
        type: integer
      id:
        description: 消息ID
        type: integer
      msg_content:
        description: 消息内容 文件则为url，文本则为文本内容，由拓展信息进行补充
        type: string
      msg_extend:
        allOf:
        - $ref: '#/definitions/model.MsgExtend'
        description: 消息扩展信息 可能为null
      msg_type:
        description: 消息类型 [text,file]
        type: string
      notify_type:
        description: 通知类型 [system,common]
        type: string
      relation_id:
        description: 关系ID
        type: integer
    type: object
  reply.ParamCreateFileMsg:
    properties:
      create_at:
        description: 创建时间
        type: string
      file_id:
        description: 文件 ID
        type: integer
      id:
        description: 消息 ID
        type: integer
      msg_content:
        description: 消息内容，文件则为 url
        type: string
    type: object
...
  reply.ParamGetMsgsByContent:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.ParamBriefMsgInfo'
        type: array
      total:
        type: integer
    type: object
  reply.ParamGetMsgsRelationIDAndTime:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.ParamMsgInfoWithRly'
        type: array
      total:
        type: integer
    type: object
...
  reply.ParamGetPinMsgsByRelationID:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.ParamMsgInfo'
        type: array
      total:
        type: integer
    type: object
...
  reply.ParamGetRlyMsgsInfoByMsgID:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.ParamMsgInfo'
        type: array
      total:
        type: integer
    type: object
...
  reply.ParamGetTopMsgByRelationID:
    properties:
      msg_info:
        allOf:
        - $ref: '#/definitions/reply.ParamMsgInfo'
        description: 置顶消息详情
    type: object
...
  reply.ParamMsgInfo:
    properties:
      account_id:
        description: 账号 ID，发送者的 ID
        type: integer
      create_at:
        description: 创建时间
        type: string
      file_id:
        description: 文件 ID，当消息类型为 file 时 > 0
        type: integer
      id:
        description: 消息 ID
        type: integer
      is_pin:
        description: 是否 pin
        type: boolean
      is_revoke:
        description: 是否撤回
        type: boolean
      is_top:
        description: 是否置顶
        type: boolean
      msg_content:
        description: 消息内容 文件则为 url，文本则为文本内容，由拓展信息进行补充
        type: string
      msg_extend:
        allOf:
        - $ref: '#/definitions/model.MsgExtend'
        description: 消息扩展信息，可能为 null
      msg_type:
        description: 消息类型 [text, file]
        type: string
      notify_type:
        description: 通知类型 [system,common]
        type: string
      pin_time:
        description: pin 时间
        type: string
      read_ids:
        description: 已读的账号ID，当请求者为发送者时为空
        items:
          type: integer
        type: array
      relation_id:
        description: 关系 ID
        type: integer
      reply_count:
        description: 回复数
        type: integer
    type: object
  reply.ParamMsgInfoWithRly:
    properties:
      account_id:
        description: 账号 ID，发送者的 ID
        type: integer
      create_at:
        description: 创建时间
        type: string
      file_id:
        description: 文件 ID，当消息类型为 file 时 > 0
        type: integer
      id:
        description: 消息 ID
        type: integer
      is_pin:
        description: 是否 pin
        type: boolean
      is_revoke:
        description: 是否撤回
        type: boolean
      is_top:
        description: 是否置顶
        type: boolean
      msg_content:
        description: 消息内容 文件则为 url，文本则为文本内容，由拓展信息进行补充
        type: string
      msg_extend:
        allOf:
        - $ref: '#/definitions/model.MsgExtend'
        description: 消息扩展信息，可能为 null
      msg_type:
        description: 消息类型 [text, file]
        type: string
      notify_type:
        description: 通知类型 [system,common]
        type: string
      pin_time:
        description: pin 时间
        type: string
      read_ids:
        description: 已读的账号ID，当请求者为发送者时为空
        items:
          type: integer
        type: array
      relation_id:
        description: 关系 ID
        type: integer
      reply_count:
        description: 回复数
        type: integer
      rly_msg:
        allOf:
        - $ref: '#/definitions/reply.ParamRlyMsg'
        description: 回复消息详情，可能为 nil
    type: object
  reply.ParamMsgInfoWithRlyAndHasRead:
    properties:
      account_id:
        description: 账号 ID，发送者的 ID
        type: integer
      create_at:
        description: 创建时间
        type: string
      file_id:
        description: 文件 ID，当消息类型为 file 时 > 0
        type: integer
      has_read:
        description: 是否已读
        type: boolean
      id:
        description: 消息 ID
        type: integer
      is_pin:
        description: 是否 pin
        type: boolean
      is_revoke:
        description: 是否撤回
        type: boolean
      is_top:
        description: 是否置顶
        type: boolean
      msg_content:
        description: 消息内容 文件则为 url，文本则为文本内容，由拓展信息进行补充
        type: string
      msg_extend:
        allOf:
        - $ref: '#/definitions/model.MsgExtend'
        description: 消息扩展信息，可能为 null
      msg_type:
        description: 消息类型 [text, file]
        type: string
      notify_type:
        description: 通知类型 [system,common]
        type: string
      pin_time:
        description: pin 时间
        type: string
      read_ids:
        description: 已读的账号ID，当请求者为发送者时为空
        items:
          type: integer
        type: array
      relation_id:
        description: 关系 ID
        type: integer
      reply_count:
        description: 回复数
        type: integer
      rly_msg:
        allOf:
        - $ref: '#/definitions/reply.ParamRlyMsg'
        description: 回复消息详情，可能为 nil
    type: object
  reply.ParamOfferMsgsByAccountIDAndTime:
    properties:
      list:
        items:
          $ref: '#/definitions/reply.ParamMsgInfoWithRlyAndHasRead'
        type: array
      total:
        type: integer
    type: object
...
  reply.ParamRlyMsg:
    properties:
      is_revoked:
        description: 是否撤回
        type: boolean
      msg_content:
        description: 消息内容
        type: string
      msg_extend:
        allOf:
        - $ref: '#/definitions/model.MsgExtend'
        description: 消息扩展信息(被@的信息)，可能为 null
      msg_id:
        description: 回复消息 ID
        type: integer
      msg_type:
        description: 消息类型[text,file]
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/message/file:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 关系id
        in: formData
        name: relation_id
        required: true
        type: integer
      - description: 回复消息id
        in: formData
        name: rly_msg_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamCreateFileMsg'
              type: object
      security:
      - BasicAuth: []
      summary: 发布文件类型的消息
      tags:
      - message
  /api/message/info/top:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情,但不存在则为null
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamGetTopMsgByRelationID'
              type: object
      summary: 获取指定关系中的置顶消息，如果不存在则为 null
      tags:
      - message
  /api/message/list/content:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: content
        required: true
        type: string
      - description: 第几页
        in: query
        name: page
        type: integer
      - description: 每页的大小
        in: query
        name: page_size
        type: integer
      - in: query
        name: relation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情，包含回复消息
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamGetMsgsByContent'
              type: object
      summary: 通过内容模糊查找指定或所有关系中的消息，按照时间先后顺序倒序排序，不会查询撤回的消息（指定关系 ID < 0 则查询所有关系中的消息）
      tags:
      - message
  /api/message/list/offer:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 拉取消息的最晚时间戳（精确到秒）
        in: query
        minimum: 1
        name: last_time
        required: true
        type: integer
      - description: 第几页
        in: query
        name: page
        type: integer
      - description: 每页的大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情，包含回复消息
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamOfferMsgsByAccountIDAndTime'
              type: object
      summary: 获取所有关系指定时间戳之后的信息，获取的消息按照发布时间先后排序，同时包含是否已读的标识
      tags:
      - message
  /api/message/list/pin:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 第几页
        in: query
        name: page
        type: integer
      - description: 每页的大小
        in: query
        name: page_size
        type: integer
      - in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情，包含回复消息
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamGetPinMsgsByRelationID'
              type: object
      summary: 获取指定关系的 pin 消息，按照pin时间倒序排序
      tags:
      - message
  /api/message/list/reply:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        minimum: 1
        name: msg_id
        required: true
        type: integer
      - description: 第几页
        in: query
        name: page
        type: integer
      - description: 每页的大小
        in: query
        name: page_size
        type: integer
      - in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情，包含回复消息
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamGetRlyMsgsInfoByMsgID'
              type: object
      summary: 获取指定消息的所有回复消息，按照回复时间先后排序
      tags:
      - message
  /api/message/list/time:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 拉取消息最晚的时间戳（精确到秒）
        in: query
        minimum: 1
        name: last_time
        required: true
        type: integer
      - description: 第几页
        in: query
        name: page
        type: integer
      - description: 每页的大小
        in: query
        name: page_size
        type: integer
      - description: 关系 ID
        in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 完整的消息详情，包含回复消息
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.ParamGetMsgsRelationIDAndTime'
              type: object
      summary: 获取指定关系指定时间戳之前的信息，获取的消息按照发布时间先后排序
      tags:
      - message
  /api/message/update/pin:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息 ID
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 是否 pin
        in: query
        name: is_pin
        required: true
        type: boolean
      - description: 关系 ID
        in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在
            5001:消息不存在
          schema:
            $ref: '#/definitions/common.State'
      summary: 更改消息的 pin 状态
      tags:
      - message
  /api/message/update/revoke:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息 ID
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在
            5001:消息不存在
          schema:
            $ref: '#/definitions/common.State'
      summary: 撤回消息
      tags:
      - message
  /api/message/update/top:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 账户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 消息 ID
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 是否置顶
        in: query
        name: is_top
        required: true
        type: boolean
      - description: 关系 ID
        in: query
        minimum: 1
        name: relation_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 2010:账号不存在
            5001:消息不存在
          schema:
            $ref: '#/definitions/common.State'
      summary: 更改消息的置顶状态
      tags:
      - message
...
swagger: "2.0"